requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.3.9
aniso8601==8.0.0
argon2-cffi==20.1.0
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bencode-python3==1.0.2
bencodepy==0.9.5
bleach==3.1.5
CacheControl==0.12.6
cachetools==4.1.0
cchardet==2.1.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
elasticsearch==7.8.0
email-validator==1.1.1
entrypoints==0.3
et-xmlfile==1.0.1
face-recognition-models==0.3.0
fake-useragent==0.1.11
firebase-admin==4.3.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
geographiclib==1.50
geopy==2.0.0
GetOldTweets3==0.0.11
gevent-websocket==0.10.1
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
googletransx==2.4.2
greenlet==0.4.16
grpcio==1.30.0
httplib2==0.18.1
hyperlink==19.0.0
idna==2.9
imutils==0.5.3
incremental==17.5.0
ipydatetime==1.1.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itemadapter==0.1.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mod-wsgi==4.7.1
monotonic==1.5
msgpack==1.0.0
multidict==4.7.6
mysql-connector-python==8.0.21
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.3.0.36
openpyxl==3.0.4
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parsel==1.6.0
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
pprint==0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.6
Protego==0.1.16
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.5.2
pymongo==3.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyquery==1.4.1
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==3.13.1
python-firebase==1.2
python-socketio==4.6.0
pytube==9.6.0
pytube3==9.6.4
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
queuelib==1.5.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
schedule==0.6.0
scikit-learn==0.23.1
scipy==1.5.1
Scrapy==2.2.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tinydb==4.1.1
tkcalendar==1.6.1
toml==0.10.1
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
tweepy==3.8.0
twint==2.1.20
Twisted==20.3.0
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.1
xlrd==1.2.0
yarl==1.4.2
zope.event==4.4
zope.interface==5.1.0
Flask-RESTful